Rename rule BA4001 to ReportPECompilerData (#386)
